relocated SideBar to components folder, edited Logs and Stats imports in App.tsx
